protocol: update tests to validate blob res
since signature is now verified on blob res rather than update,
the tests needed to be updated to handle this. Mostly this is just
a matter of setting up the header for the name, so that the syncer
can use it to fetch the pubkey and match the signature for the payload.

If anything breaks in sync, the most likely culprit is the last couple
of commits that rework the syncing mechanism to switch validating
updates to validating blob res.